# changefeedccl: fast path csv encoding when escaping isn't necessary