--- a/pkg/ccl/changefeedccl/encoder_csv.go
+++ b/pkg/ccl/changefeedccl/encoder_csv.go
@@ -15,6 +15,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/ccl/changefeedccl/cdcevent"
 	"github.com/cockroachdb/cockroach/pkg/ccl/changefeedccl/changefeedbase"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
+	"github.com/cockroachdb/cockroach/pkg/sql/types"
 	"github.com/cockroachdb/cockroach/pkg/util/encoding/csv"
 	"github.com/cockroachdb/cockroach/pkg/util/hlc"
 	"github.com/cockroachdb/errors"
@@ -55,7 +56,7 @@ func (e *csvEncoder) EncodeValue(
 	if err := updatedRow.ForEachColumn().Datum(func(d tree.Datum, col cdcevent.ResultColumn) error {
 		e.formatter.Reset()
 		e.formatter.FormatNode(d)
-		return e.writer.WriteField(&e.formatter.Buffer)
+		return e.writer.WriteField(&e.formatter.Buffer, mayNeedCSVEscaping(d.ResolvedType().Family()))
 	}); err != nil {
 		return nil, err
 	}
@@ -65,6 +66,23 @@ func (e *csvEncoder) EncodeValue(
 	}
 	e.writer.Flush()
 	return e.buf.Bytes(), nil
+}
+
+// mayNeedCSVEscaping checks whether the given type is known to encode to a charset that can't contain
+// special characters. For this encoder, special characters are double quotes, commas, leading spaces,
+// and the special 2-byte string "\.".
+func mayNeedCSVEscaping(f types.Family) bool {
+	switch f {
+	case types.BoolFamily, types.IntFamily, types.DecimalFamily, types.FloatFamily, types.BitFamily:
+		// Literals with a limited charset.
+		return false
+	case types.BytesFamily:
+		// Bytes are hex-encoded.
+		return false
+	}
+	// There are definitely more types that could return false here, but probability of error goes up
+	// with each one added so let's add them as needed.
+	return true
 }
 
 // EncodeResolvedTimestamp implements the Encoder interface.