BYTES were exporting slower than STRINGS due to their more verbose encoding. However, the nice thing about hex encoding is we know it can't contain special characters. So this commit adds a type check to see if we can safely skip the escape step, and now BYTES are faster.

name                          old time/op    new time/op    delta
CSVEncodeWideRow-16             81.1µs ± 2%    85.1µs ± 2%   +4.98%  (p=0.000 n=10+8)
CSVEncodeWideRowASCII-16        63.5µs ± 3%    66.1µs ± 3%   +4.11%  (p=0.000 n=10+10)
CSVEncodeWideColumnsASCII-16    20.0µs ± 2%    20.2µs ± 5%     ~     (p=0.549 n=10+9)
CSVEncodeWideColumns-16         31.9µs ± 1%    32.0µs ± 1%     ~     (p=0.436 n=10+10)
CSVEncodeWideColumnsBytes-16    95.7µs ± 1%    29.3µs ± 5%  -69.42%  (p=0.000 n=8+9)

name                          old alloc/op   new alloc/op   delta
CSVEncodeWideRow-16               144B ± 1%      155B ±13%   +7.44%  (p=0.000 n=8+8)
CSVEncodeWideRowASCII-16          112B ± 2%      117B ± 3%   +4.54%  (p=0.000 n=10+10)
CSVEncodeWideColumnsASCII-16     35.0B ± 0%     35.9B ± 3%     ~     (p=0.106 n=8+9)
CSVEncodeWideColumns-16          57.7B ± 4%     57.0B ± 0%     ~     (p=0.309 n=10+7)
CSVEncodeWideColumnsBytes-16      174B ± 3%       53B ± 3%  -69.29%  (p=0.000 n=10+8)

name                          old allocs/op  new allocs/op  delta
CSVEncodeWideRow-16               0.00           0.00          ~     (all equal)
CSVEncodeWideRowASCII-16          0.00           0.00          ~     (all equal)
CSVEncodeWideColumnsASCII-16      0.00           0.00          ~     (all equal)
CSVEncodeWideColumns-16           0.00           0.00          ~     (all equal)
CSVEncodeWideColumnsBytes-16      0.00           0.00          ~     (all equal)

Release note: None